requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-Script==2.0.6
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
pycodestyle==2.12.1
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
